+++ packages/health-computer/src/lib.rs
@@ -1,8 +1,10 @@
 mod data_types;
 mod health_computer;
-pub use self::{data_types::*, health_computer::*};
-
 #[cfg(feature = "javascript")]
 mod javascript;
+#[cfg(test)]
+mod tests;
+
 #[cfg(feature = "javascript")]
 pub use self::javascript::*;
+pub use self::{data_types::*, health_computer::*};
