+++ packages/health-computer/src/tests/helpers/mock_coin_info.rs
