+++ packages/health-computer/src/tests/helpers/mock_vault_config.rs
