+++ packages/health-computer/src/tests/max_borrow_vault.rs
@@ -1,19 +1,18 @@
 use std::collections::HashMap;
 
 use cosmwasm_std::{coin, Uint128};
-use mars_rover_health_computer::{DenomsData, HealthComputer, VaultsData};
 use mars_types::{
     credit_manager::Positions,
     health::{AccountKind, BorrowTarget},
 };
 
-use crate::helpers::{udai_info, umars_info};
-
-pub mod helpers;
+use super::helpers::{osmo_atom_1_config, udai_info, umars_info};
+use crate::{DenomsData, HealthComputer, VaultsData};
 
 #[test]
-fn max_borrow_wallet_offset_good() {
+fn max_borrow_vault_offset_good() {
     let udai = udai_info();
+    let osmo_atom_1_config = osmo_atom_1_config();
 
     let denoms_data = DenomsData {
         prices: HashMap::from([(udai.denom.clone(), udai.price)]),
@@ -22,7 +21,10 @@ fn max_borrow_wallet_offset_good() {
 
     let vaults_data = VaultsData {
         vault_values: Default::default(),
-        vault_configs: Default::default(),
+        vault_configs: HashMap::from([(
+            osmo_atom_1_config.addr.clone(),
+            osmo_atom_1_config.clone(),
+        )]),
     };
 
     let h = HealthComputer {
@@ -38,14 +40,22 @@ fn max_borrow_wallet_offset_good() {
         vaults_data,
     };
 
-    let max_borrow_amount =
-        h.max_borrow_amount_estimate(&udai.denom, &BorrowTarget::Wallet).unwrap();
-    assert_eq!(Uint128::new(1014), max_borrow_amount);
+    let max_borrow_amount = h
+        .max_borrow_amount_estimate(
+            &udai.denom,
+            &BorrowTarget::Vault {
+                address: osmo_atom_1_config.addr,
+            },
+        )
+        .unwrap();
+
+    assert_eq!(Uint128::new(3381), max_borrow_amount);
 }
 
 #[test]
-fn max_borrow_wallet_offset_margin_of_error() {
+fn max_borrow_vault_offset_margin_of_error() {
     let umars = umars_info();
+    let osmo_atom_1_config = osmo_atom_1_config();
 
     let denoms_data = DenomsData {
         prices: HashMap::from([(umars.denom.clone(), umars.price)]),
@@ -54,7 +64,10 @@ fn max_borrow_wallet_offset_margin_of_error() {
 
     let vaults_data = VaultsData {
         vault_values: Default::default(),
-        vault_configs: Default::default(),
+        vault_configs: HashMap::from([(
+            osmo_atom_1_config.addr.clone(),
+            osmo_atom_1_config.clone(),
+        )]),
     };
 
     let h = HealthComputer {
@@ -70,9 +83,15 @@ fn max_borrow_wallet_offset_margin_of_error() {
         vaults_data,
     };
 
-    let max_borrow_amount =
-        h.max_borrow_amount_estimate(&umars.denom, &BorrowTarget::Wallet).unwrap();
+    let max_borrow_amount = h
+        .max_borrow_amount_estimate(
+            &umars.denom,
+            &BorrowTarget::Vault {
+                address: osmo_atom_1_config.addr,
+            },
+        )
+        .unwrap();
 
-    // Normally could be 960, but conservative offset rounding has a margin of error
-    assert_eq!(Uint128::new(959), max_borrow_amount);
+    // Normally could be 3200, but conservative offset rounding has a margin of error
+    assert_eq!(Uint128::new(3196), max_borrow_amount);
 }
