+++ packages/health-computer/src/tests/max_swap_prop.rs
@@ -1,7 +1,6 @@
-use helpers::max_swap_prop_test_runner;
 use mars_types::health::SwapKind;
 
-pub mod helpers;
+use super::helpers::max_swap_prop_test_runner;
 
 #[test]
 fn max_swap_amount_default_renders_healthy_max_ltv() {
