+++ packages/health-computer/src/tests/mod.rs
@@ -1,0 +1,14 @@
+mod health_scenarios;
+mod helpers;
+mod hls;
+mod input_validation;
+mod max_borrow_deposit;
+mod max_borrow_prop;
+mod max_borrow_validation;
+mod max_borrow_vault;
+mod max_borrow_wallet;
+mod max_swap;
+mod max_swap_prop;
+mod max_swap_validation;
+mod max_withdraw;
+mod max_withdraw_prop_test;
