--- packages/health-computer/tests/helpers/mod.rs
@@ -1,10 +1,0 @@
-pub use self::{
-    mock_coin_info::*, mock_vault_config::*, prop_test_runner_borrow::*, prop_test_runner_swap::*,
-    prop_test_strategies::*,
-};
-
-mod mock_coin_info;
-mod mock_vault_config;
-mod prop_test_runner_borrow;
-mod prop_test_runner_swap;
-mod prop_test_strategies;
